Create tip calculator app
This commit implements a tip calculator app. This app has the basic
functionality of calculating a tip amount and a total amount based on a
bill amount and a tip percentage. It also has further features to change
the currency, change the color mode, split the bill among more than one
person, change the default tip percentage upon opening the app, and
remembering the bill amount for ten minutes after closing the app.